edit/delete/add img to post (user)
UI update:
- web logo
- web icon
- admin user now has red text

+++ public/likemanager.php
@@ -1,43 +1,51 @@
 <?php
     session_start();
     include 'includes/db.php';
-    if(isset($_POST['id'])) {
-        if(isset($_POST['addlike']) && isset($_POST['addlike']) == true){
-            // check whether the user has liked the post
-            if(checkLiked($conn) == false) { 
-                //insert new like into DB
-                $sqlInsert = "INSERT INTO likes (like_author_ID, like_post_ID) VALUES (?,?)";
+    include 'func/likeFunc.php';
+
+    //check signed in user
+    if(checkSignedInOnLike()) {
+        if(isset($_POST['id'])) {
+            if(isset($_POST['addlike']) && isset($_POST['addlike']) == true){
+                // check whether the user has liked the post
+                if(checkLiked($conn) == false) { 
+                    //insert new like into DB
+                    $sqlInsert = "INSERT INTO likes (like_author_ID, like_post_ID) VALUES (?,?)";
+                    $stmtInsert = $conn->prepare($sqlInsert);
+                    $stmtInsert->bind_param("ii", $_SESSION['userID'], $_POST['id']);
+                    $stmtInsert->execute();
+
+                    //update total likes of post
+                    updateTotal($conn, "add");
+                }  
+                else {
+                    $_SESSION['notification'] = "You already liked this post";
+                    $likeError['post_no_likes'] = "error";
+                    echo json_encode($likeError);
+                }
+            } 
+
+            else if(isset($_POST['removelike']) && isset($_POST['removelike']) == true) {
+                //remove like from DB
+                $sqlInsert = "DELETE FROM likes WHERE like_author_ID = ? AND like_post_ID = ?";
                 $stmtInsert = $conn->prepare($sqlInsert);
                 $stmtInsert->bind_param("ii", $_SESSION['userID'], $_POST['id']);
                 $stmtInsert->execute();
 
                 //update total likes of post
-                updateTotal($conn, "add");
-            }  
-            else {
-                $_SESSION['notification'] = "You already liked this post";
-                $likeError['post_no_likes'] = "error";
-                echo json_encode($likeError);
-            }
-        } 
-
-        else if(isset($_POST['removelike']) && isset($_POST['removelike']) == true) {
-            //remove like from DB
-            $sqlInsert = "DELETE FROM likes WHERE like_author_ID = ? AND like_post_ID = ?";
-            $stmtInsert = $conn->prepare($sqlInsert);
-            $stmtInsert->bind_param("ii", $_SESSION['userID'], $_POST['id']);
-            $stmtInsert->execute();
-
-            //update total likes of post
-            if($stmtInsert->affected_rows >= 1) {
-                updateTotal($conn, "remove"); 
-            }
-            else {
-                $_SESSION['notification'] = "You already unliked this post";
-                $likeError['post_no_likes'] = "error";
-                echo json_encode($likeError);
+                if($stmtInsert->affected_rows >= 1) {
+                    updateTotal($conn, "remove"); 
+                }
+                else {
+                    $_SESSION['notification'] = "You already unliked this post";
+                    $likeError['post_no_likes'] = "error";
+                    echo json_encode($likeError);
+                }
             }
         }
+    } else {
+        $likeError['post_no_likes'] = "error";
+        echo json_encode($likeError);
     }
 
     function checkLiked($conn) {
